removed delete attachments from main page, and added Bound To display

+++ ui/src/AttachmentsList.svelte
@@ -10,6 +10,7 @@
   const dispatch = createEventDispatcher()
 
   export let attachments: Array<HrlB64WithContext>
+  export let allowDelete = true
 
   const { getStore } :any = getContext("gzStore");
   let store: GamezStore = getStore();
@@ -17,7 +18,10 @@
 </script>
 <div class="attachments-list">
   {#each attachments as attachment, index}
-    <div class="attachment-item">
+    <div 
+      class:attachment-item-with-delete={allowDelete}
+      class:attachment-item={!allowDelete}
+    >
       {#await store.weClient.attachableInfo(hrlB64WithContextToRaw(attachment))}
         <sl-button size="small" loading></sl-button>
       {:then { attachableInfo }}
@@ -29,13 +33,15 @@
           style="display:flex;flex-direction:row;margin-right:5px"><sl-icon src={attachableInfo.icon_src} slot="prefix"></sl-icon>
           {attachableInfo.name}
         </sl-button> 
-        <sl-button size="small"
-          on:click={()=>{
-            dispatch("remove-attachment",index)
-          }}
-        >
-          <Fa icon={faTrash} />
-        </sl-button>
+        {#if allowDelete}
+          <sl-button size="small"
+            on:click={()=>{
+              dispatch("remove-attachment",index)
+            }}
+          >
+            <Fa icon={faTrash} />
+          </sl-button>
+        {/if}
       {:catch error}
         Oops. something's wrong.
       {/await}
@@ -44,13 +50,13 @@
 </div>
 <style>
   .attachments-list {
-    margin-top:5px; 
     display:flex;
     flex-direction:row;
     flex-wrap: wrap;
-
   }
   .attachment-item {
+  }
+  .attachment-item-with-delete {
     border:1px solid #aaa; 
     background-color:rgba(0,255,0,.1); 
     padding:4px;
